requirements.txt
asgiref==3.4.1
Django==3.2
djangorestframework==3.13.1
pip==20.0.2
PyJWT==1.7.1
pytz==2021.3
setuptools==46.1.1
sqlparse==0.4.2
typing-extensions==4.0.1
wheel==0.34.2
